backend-offline/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
asyncpg==0.28.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.3.0
email-validator==2.0.0
sse-starlette==1.6.5
aiofiles==23.2.1
python-dotenv==1.0.0
